fix warnings and appveyor again

+++ source/backend/llvm/translator.cpp
@@ -2358,7 +2358,7 @@ namespace backend
 							auto ut = inst->operands[0]->getType()->toUnionType();
 							auto vid = dcast(fir::ConstantInt, inst->operands[1])->getSignedValue();
 
-							iceAssert(vid < ut->getVariantCount());
+							iceAssert((size_t) vid < ut->getVariantCount());
 							auto vt = ut->getVariant(vid)->getInteriorType();
 
 							auto lut = typeToLlvm(ut, module);
